@@ -10,30 +10,36 @@
         "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
     },
     "devDependencies": {
-        "@mdi/font": "^5.1.45",
+        "@mdi/font": "^5.4.55",
         "axios": "^0.19.2",
-        "bootstrap": "^4.0.0",
+        "bootstrap": "^4.5.2",
+        "browser-sync": "^2.26.12",
+        "browser-sync-webpack-plugin": "^2.0.1",
         "cross-env": "^7.0",
+        "fibers": "^5.0.0",
         "jquery": "^3.2",
         "laravel-mix": "^5.0.1",
-        "lodash": "^4.17.13",
+        "lodash": "^4.17.19",
         "popper.js": "^1.12",
         "resolve-url-loader": "^2.3.1",
-        "sass": "^1.20.1",
-        "sass-loader": "^8.0.0",
-        "vue": "^2.5.17",
-        "vue-template-compiler": "^2.6.10"
+        "sass": "^1.26.10",
+        "sass-loader": "^9.0.0",
+        "vue": "^2.6.11",
+        "vue-template-compiler": "^2.6.10",
+        "vuetify-loader": "^1.6.0",
+        "vuetifyjs-mix-extension": "0.0.11"
     },
     "dependencies": {
         "@websanova/vue-auth": "^3.2.1-beta",
         "es6-promise": "^4.2.8",
+        "laravel-mix-alias": "^1.0.2",
         "moment": "^2.27.0",
-        "vee-validate": "^3.3.0",
+        "vee-validate": "^3.3.8",
         "vue-axios": "^2.1.5",
-        "vue-headful": "^2.1.0",
+        "vue-meta": "^2.4.0",
         "vue-moment": "^4.1.0",
-        "vue-router": "^3.3.1",
-        "vuetify": "^2.2.27",
-        "vuex": "^3.3.0"
+        "vue-router": "^3.4.2",
+        "vuetify": "^2.3.8",
+        "vuex": "^3.5.1"
     }
 }
