@@ -5,29 +5,26 @@
       indeterminate
       absolute
       top
-      v-bind:class="{ loader: true }"
     ></v-progress-linear>
     <v-overlay
       :value="80"
       style="z-index: 102"
-    >
-      <v-progress-circular
+    > 
+       <v-progress-circular
         style="z-index: 103"
-        :size="100"
+        :size="70"
         :width="7"
         color="white"
         indeterminate
       ></v-progress-circular>
     </v-overlay>
-
   </div>
 </template>
 
 <script>
 export default {
-  data () {
-    return {
-    }
-  }
+  name: 'loader',
 }
 </script>
+
+<style></style>
